+++ tests/table/apis/table.entity.query.test.ts
@@ -1004,6 +1004,8 @@ describe("table Entity APIs test - using Azure/data-tables", () => {
     await tableClient.deleteTable();
   });
 
+  // issue 1828
+
   it("16. should find guids when using filter with ge, lt, gt and ne, @loki", async () => {
     const tableClient = createAzureDataTablesClient(
       testLocalAzuriteInstance,
@@ -1113,6 +1115,57 @@ describe("table Entity APIs test - using Azure/data-tables", () => {
       testsCompleted++;
     }
     assert.strictEqual(testsCompleted, queriesAndExpectedResult.length);
+    await tableClient.deleteTable();
+  });
+
+  it("17. should work correctly when query filter single boolean and partition filter, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("partwithnot")
+    );
+
+    await tableClient.createTable();
+
+    const result1 = await tableClient.createEntity({
+      partitionKey: `Part1`,
+      rowKey: `1`,
+      foo: "testEntity1"
+    });
+    assert.notStrictEqual(result1.etag, undefined);
+
+    const result2 = await tableClient.createEntity({
+      partitionKey: `Part2`,
+      rowKey: `1`,
+      foo: "testEntity2"
+    });
+    assert.notStrictEqual(result2.etag, undefined);
+
+    const result3 = await tableClient.createEntity({
+      partitionKey: `Part3`,
+      rowKey: `1`,
+      foo: "testEntity3"
+    });
+    assert.notStrictEqual(result3.etag, undefined);
+
+    const maxPageSize = 5;
+    const entities = tableClient.listEntities<TableEntity<{ foo: string }>>({
+      queryOptions: {
+        filter: odata`not (PartitionKey lt 'Part2')`
+      }
+    });
+    let all: TableEntity<{ foo: string }>[] = [];
+    for await (const entity of entities.byPage({
+      maxPageSize
+    })) {
+      all = [...all, ...entity];
+    }
+    assert.strictEqual(all.length, 2);
+    all.forEach((entity) => {
+      assert.ok(
+        entity.partitionKey === `Part2` || entity.partitionKey === `Part3`
+      );
+    });
+
     await tableClient.deleteTable();
   });
 });