+++ tests/table/unit/LokiJsQueryTranscriber.unit.test.2.ts
@@ -3,41 +3,63 @@ import LokiJsQueryTranscriberFactory from "../../../src/table/persistence/QueryT
 
 const entityQueries = [
   {
+    input: "PartitionKey eq 'azurite' and RowKey eq 'tables'",
+    expected:
+      "return ( item.properties.PartitionKey === `azurite` && item.properties.RowKey === `tables` )"
+  },
+  {
+    input: "PartitionKey eq 'azurite' or RowKey eq 'tables'",
+    expected:
+      "return ( item.properties.PartitionKey === `azurite` || item.properties.RowKey === `tables` )"
+  },
+  {
+    input: "PartitionKey eq 'Foo '''' Bar'",
+    expected: "return ( item.properties.PartitionKey === `Foo '' Bar` )"
+  },
+  {
+    input: "PartitionKey eq 'Foo '' Bar'",
+    expected: "return ( item.properties.PartitionKey === `Foo ' Bar` )"
+  },
+  {
+    input: "not (PartitionKey lt 'Part2')",
+    expected: "return ( ! ( item.properties.PartitionKey < `Part2` ) )"
+  },
+  {
     input: "PartitionKey eq 'azurite'",
-    expected: "return ( item.PartitionKey === `azurite` )"
+    expected: "return ( item.properties.PartitionKey === `azurite` )"
   },
   {
     input: "RowKey eq 'azurite'",
-    expected: "return ( item.RowKey === `azurite` )"
+    expected: "return ( item.properties.RowKey === `azurite` )"
   },
   {
     input: "PartitionKey gt 'azurite'",
-    expected: "return ( item.PartitionKey > `azurite` )"
+    expected: "return ( item.properties.PartitionKey > `azurite` )"
   },
   {
     input: "PartitionKey ge 'azurite'",
-    expected: "return ( item.PartitionKey >= `azurite` )"
+    expected: "return ( item.properties.PartitionKey >= `azurite` )"
   },
   {
     input: "PartitionKey lt 'azurite'",
-    expected: "return ( item.PartitionKey < `azurite` )"
+    expected: "return ( item.properties.PartitionKey < `azurite` )"
   },
   {
     input: "PartitionKey le 'azurite'",
-    expected: "return ( item.PartitionKey <= `azurite` )"
+    expected: "return ( item.properties.PartitionKey <= `azurite` )"
   },
   {
     input: "PartitionKey ne 'azurite'",
-    expected: "return ( item.PartitionKey !== `azurite` )"
+    expected: "return ( item.properties.PartitionKey !== `azurite` )"
   },
   {
     input: "not (PartitionKey eq 'azurite')",
-    expected: "return ( ! ( item.PartitionKey === `azurite` ) )"
+    expected: "return ( ! ( item.properties.PartitionKey === `azurite` ) )"
   },
   {
     input: "MyField gt datetime'2021-06-05T16:20:00'",
     expected:
-      "return ( new Date(item.properties.MyField).getTime() > new Date(`2021-06-05T16:20:00`).getTime() )"
+      "return ( new Date(item.properties.MyField).getTime() > new Date('2021-06-05T16:20:00').getTime() )"
   },
   {
     input: "MyField gt 1337",
@@ -48,27 +70,18 @@ const entityQueries = [
     expected: "return ( item.properties.MyField > '1337' )"
   },
   {
-    input: "PartitionKey eq 'azurite' and RowKey eq 'tables'",
-    expected:
-      "return ( item.PartitionKey === `azurite` && item.RowKey === `tables` )"
-  },
-  {
-    input: "PartitionKey eq 'azurite' or RowKey eq 'tables'",
-    expected:
-      "return ( item.PartitionKey === `azurite` || item.RowKey === `tables` )"
-  },
-  {
     input: "MyField eq guid'00000000-0000-0000-0000-000000000000'",
     expected:
-      "return ( item.properties.MyField === `00000000-0000-0000-0000-000000000000` )"
+      "return ( ( item.properties.MyField === '00000000-0000-0000-0000-000000000000' ) || ( item.properties.MyField === 'MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw' ) )"
   },
   {
     input: "PartitionKey eq 'Iam''good''atTypeScript'",
-    expected: "return ( item.PartitionKey === `Iam'good'atTypeScript` )"
+    expected:
+      "return ( item.properties.PartitionKey === `Iam'good'atTypeScript` )"
   },
   {
     input: "PartitionKey eq 'Isn''tThisANastyPK'",
-    expected: "return ( item.PartitionKey === `Isn'tThisANastyPK` )"
+    expected: "return ( item.properties.PartitionKey === `Isn'tThisANastyPK` )"
   },
   {
     input: "1 eq 1",
@@ -76,67 +89,70 @@ const entityQueries = [
   },
   {
     input: "PartitionKey eq 'a'",
-    expected: "return ( item.PartitionKey === `a` )"
+    expected: "return ( item.properties.PartitionKey === `a` )"
   },
   {
     input: "PartitionKey eq ' '",
-    expected: "return ( item.PartitionKey === ` ` )"
+    expected: "return ( item.properties.PartitionKey === ` ` )"
   },
   {
     input: "PartitionKey eq 'Foo Bar'",
-    expected: "return ( item.PartitionKey === `Foo Bar` )"
+    expected: "return ( item.properties.PartitionKey === `Foo Bar` )"
   },
   {
     input: "PartitionKey eq 'A''Foo Bar''Z'",
-    expected: "return ( item.PartitionKey === `A'Foo Bar'Z` )"
+    expected: "return ( item.properties.PartitionKey === `A'Foo Bar'Z` )"
   },
   {
     input: "PartitionKey eq '''Foo Bar'",
-    expected: "return ( item.PartitionKey === `'Foo Bar` )"
-  },
-  {
-    input: "PartitionKey eq 'Foo '' Bar'",
-    expected: "return ( item.PartitionKey === `Foo ' Bar` )"
+    expected: "return ( item.properties.PartitionKey === `'Foo Bar` )"
   },
   {
     input: "PartitionKey eq 'Foo Bar'''",
-    expected: "return ( item.PartitionKey === `Foo Bar'` )"
+    expected: "return ( item.properties.PartitionKey === `Foo Bar'` )"
   },
   {
     input: "PartitionKey eq ' Foo Bar '",
-    expected: "return ( item.PartitionKey === ` Foo Bar ` )"
+    expected: "return ( item.properties.PartitionKey === ` Foo Bar ` )"
   },
   {
     input: "PartitionKey eq ''",
-    expected: "return ( item.PartitionKey === `` )"
+    expected: "return ( item.properties.PartitionKey === `` )"
   },
   {
     input: "PartitionKey eq '''Foo Bar'''",
-    expected: "return ( item.PartitionKey === `'Foo Bar'` )"
+    expected: "return ( item.properties.PartitionKey === `'Foo Bar'` )"
   },
   {
     input: "PartitionKey eq ''''",
-    expected: "return ( item.PartitionKey === `'` )"
+    expected: "return ( item.properties.PartitionKey === `'` )"
   },
   {
     input: "PartitionKey eq ''''''",
-    expected: "return ( item.PartitionKey === `''` )"
+    expected: "return ( item.properties.PartitionKey === `''` )"
   },
   {
     input: "PartitionKey eq ''''''''",
-    expected: "return ( item.PartitionKey === `'''` )"
+    expected: "return ( item.properties.PartitionKey === `'''` )"
   },
   {
     input: "PartitionKey eq ''''''''''",
-    expected: "return ( item.PartitionKey === `''''` )"
+    expected: "return ( item.properties.PartitionKey === `''''` )"
   },
   {
     input: "PartitionKey eq 'I am ''good'' at TypeScript'",
-    expected: "return ( item.PartitionKey === `I am 'good' at TypeScript` )"
+    expected:
+      "return ( item.properties.PartitionKey === `I am 'good' at TypeScript` )"
   },
   {
     input: "_foo eq 'bar'",
-    expected: "return ( _foo === `bar` )"
+    expected: "return ( item.properties._foo === `bar` )"
+  },
+  {
+    input:
+      "please eq 'never query ''this'' eq this or suchandsuch eq ''worse'''",
+    expected:
+      "return ( item.properties.please === `never query 'this' eq this or suchandsuch eq 'worse'` )"
   }
 ];
 
@@ -154,12 +170,15 @@ describe("Unit tests for converting an entity OData query to a JavaScript query 
         const actual = queryTranscriber.getTranscribedQuery();
         assert.strictEqual(actual, expected);
       } catch (err: any) {
-        if (input === "1 eq 1")
+        if (input === "1 eq 1") {
           assert.strictEqual(
             err.message,
             "Invalid number of terms in query!",
             `Did not get expected error on invalid query ${input}`
           );
+        } else {
+          assert.ifError(err);
+        }
       }
       done();
     });