+++ ChangeLog.md
@@ -26,6 +26,7 @@ Blob:
 Table:
 
 - Fixed issue for querying on identifiers starting with underscore.
+- Corrected query parsing logic for single boolean terms.
 - Fixed issue for querying GUIDs using operators other than eq and ne
 - GUID queries only support persistent storage on legacy (string) format GUIDs for eq and ne operators, other operators will only evaluate newly stored entities.